Client-side module for playing the game
It sends a request to the server to initialize a game of Tic-Tac-Toe. If the server responds, a socket connection is created, and the game can be played until one of the player wins. If a player enters invalid token or tries to place his token in the place of an existing token, his chance at playing is forfeited.